footer update (added tos and pp)
