+++ src/components/Site/Servers.vue
@@ -2,40 +2,22 @@
   <img src="../../assets/img/servers.top.svg" class="fill-deh-server-slider" />
   <!-- top-10 reltativ -> moves the svg down but i should jsut create a new wave -->
   <section class="w-full bg-deh-server-slider">
-    <div
-      class="font-bold text-deh-white text-center text-5xl relative lg:-top-10"
-    >
+    <div class="font-bold text-deh-white text-center text-5xl relative lg:-top-10">
       <p class="underline">{{ $t("servers.server") }}</p>
       <div class="font-normal no-underline text-2xl">
-        {{ $t("servers.desc", { c: servers }) }}
-        <span
-          class="text-deh-white bg-deh-main rounded p-[2px] text-xs align-top font-bold"
-          >BOT</span
-        >
+        {{ $t("servers.desc", { c: serverString }) }}
+        <span class="text-deh-white bg-deh-main rounded p-[2px] text-xs align-top font-bold">BOT</span>
       </div>
     </div>
     <template v-if="showDetaildServer && dummyMarqueeItems.length > 0">
       <div class="mx-8 flex flex-col items-center z-[2]">
-        <Vue3Marquee
-          :clone="true"
-          :duration="25"
-          :direction="'reverse'"
-          class="lg:max-w-7xl"
-          :gradient="true"
-          :gradient-color="[25, 25, 25]"
-          gradient-length="15%"
-        >
+        <Vue3Marquee :clone="true" :duration="25" :direction="'reverse'" class="lg:max-w-7xl" :gradient="true"
+          :gradient-color="[25, 25, 25]" gradient-length="15%">
           <div v-for="item in dummyMarqueeItems">
-            <a
-              class="bg-slate-800 p-1 m-3 rounded w-56 h-[90px] flex"
-              :href="item.invite"
-            >
-              <img
-                :src="item.img"
-                class="m-1 rounded-xl border-deh-border border-2"
-              />
+            <a class="bg-slate-800 p-1 m-3 rounded w-56 h-[90px] flex" :href="item.invite">
+              <img :src="item.img" class="m-1 rounded-xl border-deh-border border-2" />
               <div class="pt-2 text-deh-second font-semibold">
-                <div>{{ shortString(item.title)  }}</div>
+                <div>{{ shortString(item.title) }}</div>
                 <div class="font-normal text-deh-black">
                   +{{ item.members }} {{ $t('servers.members') }}
                 </div>
@@ -43,23 +25,11 @@
             </a>
           </div>
         </Vue3Marquee>
-        <Vue3Marquee
-          :clone="true"
-          :duration="25"
-          class="lg:max-w-7xl"
-          :gradient="true"
-          :gradient-color="[25, 25, 25]"
-          gradient-length="15%"
-        >
+        <Vue3Marquee :clone="true" :duration="25" class="lg:max-w-7xl" :gradient="true" :gradient-color="[25, 25, 25]"
+          gradient-length="15%">
           <div v-for="item in dummyMarqueeItems">
-            <a
-              class="bg-slate-800 p-1 m-3 rounded w-56 h-[90px] flex"
-              :href="item.invite"
-            >
-              <img
-                :src="item.img"
-                class="m-1 rounded-xl border-deh-border border-2"
-              />
+            <a class="bg-slate-800 p-1 m-3 rounded w-56 h-[90px] flex" :href="item.invite">
+              <img :src="item.img" class="m-1 rounded-xl border-deh-border border-2" />
               <div class="pt-2 text-deh-second font-semibold">
                 <div>{{ shortString(item.title) }}</div>
                 <div class="font-normal text-deh-black">
@@ -72,13 +42,12 @@
       </div>
     </template>
   </section>
-  <img
-    src="../../assets/img/servers.bottom.svg"
-    class="fill-deh-server-slider"
-  />
+  <img src="../../assets/img/servers.bottom.svg" class="fill-deh-server-slider" />
 </template>
 
 <script setup lang="ts">
+import { onBeforeMount, ref } from 'vue';
+
 type marqeeObject = {
   img: string;
   title: string;
@@ -88,7 +57,7 @@ type marqeeObject = {
   invite: string;
 };
 
-function shortString(title:string) {
+function shortString(title: string) {
   return title.length >= 15 ? (title.substring(0, 12) + "...") : title
 }
 
@@ -134,32 +103,24 @@ const dummyMarqueeItems: marqeeObject[] = [
     invite: "",
   },
 ];
+
+let serverString = "Many";
+let showDetaildServer = ref(false);
+
+function getPosts() {
+  fetch(
+    "https://japi.rest/discord/v1/application/1078340529932222505"
+  ).then(async (req) => {
+    if (req.ok) {
+      serverString = `${((await req.json()) as any).data.bot.approximate_guild_count
+        }`;
+    }
+  });
+}
+
+onBeforeMount(() => {
+  getPosts();
+})
+
 </script>
 
-<script lang="ts">
-export default {
-  data() {
-    return {
-      servers: "Many",
-      showDetaildServer: true,
-      loading: false,
-    };
-  },
-  methods: {
-    getPosts() {
-      fetch(
-        "https://japi.rest/discord/v1/application/1078340529932222505"
-      ).then(async (req) => {
-        if (req.ok) {
-          this.servers = `${
-            ((await req.json()) as any).data.bot.approximate_guild_count
-          }`;
-        }
-      });
-    },
-  },
-  beforeMount() {
-    this.getPosts();
-  },
-};
-</script>
